Include test files in default tsconfig
The tests were not included in the original tsconfig.json, only in
tsconfig-mocha.json. However, since VS Code only picks up tsconfig.json,
it did not recognized that the tests could use the decorators. 

Rename tsconfig-mocha.json to tsconfig.json, and the original
tsconfig.json to tsconfig-pack.json. That file is used for npm
publishing only.