Add JsDoc notations to all decorators and public classes
In most cases the JsDoc text is a summary of the markdown documentation.

Some JsDoc annotations need to be duplicated due to decorator working
very differently with and without parameters. Using the same JsDoc for
different signatures seems appropriate in the case of decorators, as
these would just be optional parameter in traditional functions.

Some minor rephrasing in the markdown documentation.

Remove unnecessary constructor from Injector class.