Support two-way binding on non-widget objects
This allows @bind({all: bindings]) to establish multiple two-way
bindings between the object assigned to the decorated property and the
component children. @bindAll is a shorthand for this syntax since
overloading @bind to take the bindings directly would be ambiguous.

The implementation of this feature is (and has to be) fundamentally
different from the way @bind(path) works, has the latter is really just
a re-direct. However, the perceived behavior is the same.

Add one example "bind-two-way-model" and extend the example
"bind-two-way-change-events", both linked in the documentation.

The "utils" module was getting too lengthy according to ts-lint rules
and was split in to "utils" and "utils-databinding". Module
"bind-one-way" was split in to "processOneWayBindings" and
"applyJsxBindings" to make it more similar to the modules implementing
the two-way binding.