Add selection API to ListView
The "onSelect" event issues ListViewSelectEvent instances that contain
various information about the selection event. The event itself can be
triggered simply by setting the Cell property "selectable" to true and
tapping it.

More advanced configurations are possible by attaching pre-defined
listeners (static fields on ListView) to cells or a cell child. This
also allows passing on a numeric value which is designed to indicate
different kinds of actions.