Implement ItemPicker
This is the `Picker` based counterpart to `ListView`. A `List` or
`Array` instance may be given to its `items` property and manipulated at
will. The `selectionIndex` property and `select` event are translated to
a `selection` property and `itemSelect` event. The items are translated
to displayable text either by stringification or by defining a data
binding and assigning it to the `textSource` property.

Minor list to `List` and `ListLike`: Add missing indexer.
Minor list to `ListView`: Make children optional.