Add example launch scripts for windows and linu/mac/GitPod
Both can be called via "npm start <example>" thanks to the
"run-script-os" package.

The scripts can also install a locally built tabris-decorators package
in to the example. This requires calling the script directly with the
--dev parameter.

Remove the "rm" command from the built script since it is not strictly
needed and does not work on all windows installation.

Use different tabris version for devDependencies and peerDependencies.
During development always use the latest nightly, but during production
it should work with non-nightlies.