Allow one-way bindings in .jsx files
The @property decorator and the "bind-xyz" JSX syntax already work in
JavaScript/JSX files, but if the property has no meta-data (added by
tsc) and no type guard it is marked as "unchecked". If a binding targets
such a property (as a receiver of new values) the binding would fail
with an exception. Now it only prints a warning, unless the new
"strictMode" flag on the ExtenedJSX object is set to true.

Also allow the @event decorator on properties without meta-data, just
like @property. This simple change allows it to be used in JavaScript
and has almost no effect on TypeScript code.

Further changes:
* Set strictMode to true in all TypeScript examples
* Configure tslint to include js files
* Configure unit tests to enable js/jsx spec files
* Add bind-one-way-jsx spec mirroring bind-one-way tsx spec
* Add examples for JSX one-way bindings