Make @bind JavaScript compatible
All decorators except @inject can now be used in tsc-compiled JS/JSX
files. The compiler option "experimentalDecorators" still needs to be
enabled, but "emitDecoratorMetadata" is no longer required.

In "strict" mode @bind requires explicit type information for one or
both ends of a binding via the new "type" option, or a type guard. When
not in strict mode this is optional, but an "unsafe binding" warning may
be printed. Some related error messages have been changed to mention
"type checks" instead of just "type guard". Some refactoring was
required to be able to be able to determine strict mode for any given
object (specific to the JSX processor/injector).

JavaScript/JSX specific tests for @bind were added. Some tests had to be
adjusted to always run in strict mode by default and adapt the error
messages.

Documentation will be adjusted in a later commit.