Rework two-way bindings
This is mostly a refactoring, but changes many error messages and some
edge case behaviors. This is the result of combining the implementations
of @bind(path) and @bindAll(bindings). The former was previously using a
completely different mechanism that really just forwarded setter and
getter calls and not actually synchronizing two values.

The most significant change is that the sync-back behavior of @bindAll
now also applies to @bind(path). Since the details of this were
previously not documented, this is not considered a breaking change.