Refactor @property
Extract CustomPropertyDescriptor.ts from property.ts. For each
@property call (so not per target instance) a new instance of this
class will be created and used to define the property via
Object.defineProperty. It will also be attached as property meta data
so the object can be queried later. This also opens up the possibility
of changing the configuration of the property after its initial
definition, but this is not yet implemented.

supportsChangeEvents function moved from util-databinding to subscribe,
since it's the only user and util-databinding should not gain a
dependency to CustomPropertyDescriptor.