Implement "convert" function
This function converts a given value to another type using various
strategies.

This function will be used to implement @property({convert: 'auto'}). It
will provide a certain amount of error tolerance for data binding and
dealing with "any" type variables, and of course when using JavaScript
instead of TypeScript. It will also enable users to use the tabris data
types (e.g. ColorValue) in custom components.