Add @property options "default" and "nullable"
When "nullable" is false the property can not be set to null or
undefined. Instead it will throw an error.

When "default" is set the property will have this value before it is set
for the first time. This differs from assigning a value in the property
declaration which sets the property AFTER the super() call.

When both options are combined the property will reset itself to the
default value when null or undefined are assigned to it.

Minor fix in equal.ts to handle comparison with null correctly.