Add connect function and StateProvider
This allows mapping the state of a redux store (or any implementation of
StateProvider) to any widget. Actions are not supported yet.

Store is obtained via DI, with the used injector either self-injected
when used vis JSX/inject.create, or passed explicitly as an additional
constructor parameter. If no injector is given the default injector is
used.

For the first time this adds a vanilla-js example. This required some
changes to eslintrc.json.

Documentation to follow when redux support is complete.