Add StateProvider "dispatch" function
With this change the fulltrip redux -> tabris -> redux  roundtrip is now
possible for custom components. Actions dispatcher may be attached
directly as a widget property, or by registering them as a listener. The
framework assumes that listeners are using the "on[Event]" pattern,
otherwise the distinction between the two will not work.

As with the "DefaultRootState" type, the "AnyAction" type can be
augmented to get automatic global type checking for action
creators/dispatchers without having to explicitly import a specific
type. However, due to the way declaration merging in TypeScript works
the "AnyAction" can not be augmented directly. Instead an interface
"DeaultActions" is used that defines all known actions as properties.
The name of these properties is irrelevant, the only purpose of the
interface is to serve as an augmentable pseudo-namespace.