Remove "@component" functionality from "@connect"
So far applying the "connect" decorator to a custom component would also
call "component" on it. This is convenient, but conceptionally
inconsistent. "component" modifies the original class, while "connect"
is supposed to only wrap it so it can be connected multiple times in
different ways. Also, it is feasible that a class is supposed to be
connected but not encapsulated.

Also make the text size for the different connect examples consistent.