Add option to @component to wrap constructor as a factory
Based on the new "asFactory" function of tabris 3.6. Unfortunately this
requires a separate function signature and can only be used when
component is calles as a function. This is because a class decorator can
not modify the interface of the constructor.

Use the new factory API in the connect-js example project.

Removed one (legitimately) broken test.