Allow overriding the injector used for a custom JSX element
When passing an injector to a component decorator it will be used for
all injections of that component when instanciated via JSX. By default
the injector associated with the jsx processor would be used. A custom
jsx processor with a custom injector (as opposed to the global one)
could also be used, but this would not work for the component-as-factory
feature. This new API is the only way to control which injector will be
used when creating a component via the new factory API.