Prevent @connect from double-wrapping factory components
Factories as created by asFactory or @component work via proxies, as
does @connect. To prevent a proxy-in-proxy scenario @connect will detect
factories and make a modified copies instead of wrapping. It will also
prevent any attempt to make the result a factory by denying access to
the original component.

Minor fix in ExtendedJSX to enabled test scenario with custom injector:
A component that is explicitly assigned a custom injector will always
be created with that injector, even if it doesn't have injections. This
is done so Injector.get will work on that component.

Fix two compile errors in component-factory.spec.ts