Add support for special "apply" entry in @connect mapped state objects
When adding "apply" entries the stateToProps or dispatchToProps return
values it is not treated as a property but a call to the "apply"
function. That makes it possible to interact with children of the
connected component. For class based custom components that would break
encapsulation (it won't work if @component is used), but for functional
components it's ideal.

This feature uses internal tabris API that only works with widgets,
therefore "connect" can no longer work with any object that implements
"set", which is no longer needed. For now this seem sufficient. These
changes slightly alter existing behavior and type declarations,
therefore some existing tests are adjusted.

Examples are updated and revised.