Support dependency injection on properties
So far only constructor parameters could be injected. Injecting
properties is difficult because they need to be available in the
constructor, so before the injector has obtained an instance to fulfill
the injections.

This is solved by resolving the injections lazily in a custom getter.
However, this can still be at a point in time where the injector has
not tagged the instance, meaning the injector instance must be
obtained from the prototype. Injector.get is changed to enable this.

To make sure failed property injections are discovered early the
injector forces them to be resolved (using custom metadata) once it is
idle.