Support dependency injection on properties
So far only constructor parameters could be injected. Making the
framework inject to properties is difficult because the injections need
to be already available inside the constructor of the target object. At
that time the Injector has no reference to the instance yet.

This is solved by resolving the injections lazily in a custom getter.
However, this can still be at a point in time where the injector has
not tagged the instance, meaning the injector instance must be
obtained from the prototype. Injector.get is changed to enable this.

To make sure failed property injections are discovered early the
injector forces them to be resolved (using custom metadata) once it is
idle.

This feature can help resolving circular injections. A section on this
topic is added to the documentation.