Do not set undefined source properties for one-way (to child) bindings
One edge case for two-way bindings is that a source property is
"dominant" unless it is undefined. In that specific case a target
property will be applied to the source when the binding is established.

However, this should not be the case if the binding has been marked as
one-way using the ">>" syntax.

Fix #66