Introduce value conversion feature for @bind and @bindAll
For one-way conversions the established converter function pattern is
continued. The function is given as part of the decorator configuration
object.

For two-way conversions the same converter function is tasked with
handling both directions. A second parameter is added that allows to
identify what the direction/target of the conversion is.

This object is not used in ItemPicker, which is why it can no longer
share the "Binding" interface with the other data binding code.

Add example application "bind-two-way-convert".

Minor additional fixes/adjustments in documentation and tests.

Delete one empty file.