Make @prop accept null as objects by default
It makes sense for @prop to be non-nullable for primitives as this
results in it converting null-ish values to falsy primitives.

But for objects setting "nullable" false causes exception to be thrown
for nullish values, which in practice is rarely desirable.

Since this causes the property to be less strict this is technically
not a breaking change.