Support binding listeners to model events
@bindAll can now register listeners to a model events. This can be
useful in cases where the UI is supposed to react to a signal, not a
state change.

Documentation, tests and examples are included. Tabris dependency is
temporarily set to a nightly build to ensure it works with recent
changes. It will be set to `3.8.0` when that module has been published.